refactor: Consolidate the triplicated AsyncGate/waitUntil test helpers
## Summary
- Hoist the event-driven `AsyncGate`/`waitUntil`/`TestFailure` wait primitives
  out of three drifted per-bundle copies into a single shared implementation,
  closing the maintenance hazard CLAUDE.md's "keep the copies aligned"
  mandate warned about
- The copies had already drifted: only the KernovaTests copy cancelled its
  timeout backstop `Task` on early resolution; the other two leaked it until
  deadline on every happy-path wait

## Changes
- Add a new `KernovaTestSupport` SwiftPM product/target
  (`KernovaKit/Sources/KernovaTestSupport/AsyncWaits.swift`) holding the
  single canonical `TestFailure`, `ResumeOnce`, `AsyncGate`, and `waitUntil`
  — Xcode 16 synchronized folders make the KernovaTests/KernovaMacOSAgentTests
  source folders target-private, so a shared SwiftPM product (not a shared
  file) is the mechanism that lets all three test targets depend on one copy
- `AsyncGate.wait`/`waitUntil` take the caller's isolation via
  `isolation: isolated (any Actor)? = #isolation`, eliminating the isolation
  fork the three copies previously required (`@MainActor` + non-Sendable
  predicate in KernovaTests vs. `nonisolated` + `@Sendable` elsewhere) — one
  implementation now serves every bundle
- Wire the new product into `Kernova.xcodeproj/project.pbxproj`: add
  `KernovaTestSupport` `XCSwiftPackageProductDependency`/`PBXBuildFile`
  entries and link them into the `KernovaTests` and `KernovaMacOSAgentTests`
  targets' Frameworks phases, mirroring the existing KernovaKit folder-peer
  package wiring
- Strip the duplicated `AsyncGate`/`ResumeOnce`/`waitUntil`/`TestFailure` out
  of `KernovaTests/TestHelpers.swift`, `KernovaMacOSAgentTests/TestHelpers.swift`,
  and `KernovaKit/Tests/KernovaKitTests/StreamTestSupport.swift`, replacing
  them with `import KernovaTestSupport`; `waitForChange` stays local to
  KernovaTests (it was never one of the triplicated copies), and
  `pollUntil`/`StreamTestFailure` stay local to KernovaKitTests (package-
  specific, single-copy, not part of the drift this cleans up)
- Add the same import to every other test file that referenced these symbols
  directly, and update CLAUDE.md's "Async waits in tests" table and
  ARCHITECTURE.md's KernovaKit/KernovaTests/KernovaMacOSAgentTests package
  layout to describe the new shared module

## Test plan
- [x] `make test` — all three test targets (KernovaTests, KernovaMacOSAgentTests,
      KernovaKitTests) pass via `Kernova.xctestplan`
- [x] `swift test` inside `KernovaKit/` — 239 package tests pass standalone,
      confirming the isolation-agnostic `AsyncGate` compiles and behaves
      correctly for both `@MainActor` and nonisolated callers
- [x] `make lint` — swift-format strict passes, including the new
      `AllPublicDeclarationsHaveDocumentation` requirements on the public API
- [x] `make build` — full app build succeeds

Fixes #526
