+++ KernovaKit/Tests/KernovaKitTests/ClipboardStreamTests.swift
@@ -1,6 +1,7 @@
 import CryptoKit
 import Foundation
 import Testing
+import KernovaTestSupport
 
 @testable import KernovaKit
