+++ KernovaKit/Tests/KernovaKitTests/FileProviderDomainHostTests.swift
@@ -1,6 +1,7 @@
 import FileProvider
 import Foundation
 import Testing
+import KernovaTestSupport
 
 @testable import KernovaKit
