+++ KernovaMacOSAgentTests/VsockGuestClientTests.swift
@@ -2,6 +2,7 @@ import Testing
 import Foundation
 import Darwin
 import KernovaKit
+import KernovaTestSupport
 
 @Suite("VsockGuestClient connect/retry/stop lifecycle")
 struct VsockGuestClientTests {
