+++ KernovaMacOSAgentTests/VsockGuestControlAgentTests.swift
@@ -2,6 +2,7 @@ import Testing
 import Foundation
 import Darwin
 import KernovaKit
+import KernovaTestSupport
 
 @Suite("VsockGuestControlAgent")
 struct VsockGuestControlAgentTests {
