+++ KernovaTests/ClipboardContentViewControllerRetentionTests.swift
@@ -1,6 +1,7 @@
 import AppKit
 import Foundation
 import Testing
+import KernovaTestSupport
 
 @testable import Kernova
 @testable import KernovaKit
