+++ KernovaTests/SerialSocketRelayTests.swift
@@ -1,6 +1,7 @@
 import Darwin
 import Foundation
 import Testing
+import KernovaTestSupport
 
 @testable import Kernova
