+++ KernovaTests/VsockGuestLogServiceTests.swift
@@ -2,6 +2,7 @@ import Testing
 import Foundation
 import Darwin
 import KernovaKit
+import KernovaTestSupport
 @testable import Kernova
 
 @Suite("VsockGuestLogService")
